Updated code

+++ controller/bootcamps.js
@@ -14,6 +14,8 @@ exports.getBootcamps = asyncHandler(async (req, res, next) => {
       .json(res.advancedResults);
 });
 
+
+
 // @desc Get single bootcamps
 // @route GET /api/v1/bootcamps/:id
 //access  Public
@@ -32,10 +34,23 @@ exports.getBootcamp = asyncHandler(async (req, res, next) => {
     res.status(200).json({ success: true, data: bootcamp });
 });
 
+
+
 // @desc Create new bootcamp
 // @route POST /api/v1/bootcamps/:id
 //access  Private
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
+      // Add user to req.body
+      req.body.user = req.user.id;
+
+      //check for published bootcamp
+      const publishedBootcamp = await Bootcamp.findOne({ user: req.user.id });
+
+      // if the user is not an admin, they can only add one Bootcamp
+      if (publishedBootcamp && req.user.role !== "admin") {
+        return next( new ErrorResponse(`The user with ID ${req.user.id} has already published a bootcamp`, 400));
+            }
+
     const bootcamp = await Bootcamp.create(req.body);
 
     res.status(201).json({
@@ -44,14 +59,13 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
     });
 });
 
+
+
 // @desc Update bootcamp
 // @route PUT /api/v1/bootcamps/:id
 //access  Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-    const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-      runValidators: true,
-    });
+    let bootcamp = await Bootcamp.findById(req.params.id);
 
     if (!bootcamp) {
       return next(
@@ -62,8 +76,22 @@ exports.updateBootcamp = asyncHandler(async (req, res, next) => {
       );
     }
 
+    //Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+      return next( new ErrorResponse(`User ${req.user.id} is not authorized to update this bootcamp`, 401));
+      }
+
+
+      bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+        new: true,
+        runValidators: true,
+      })
+
+
     res.status(200).json({ success: true, data: bootcamp });
 });
+
+
 
 // @desc Delete bootcamps
 // @route DELETE /api/v1/bootcamps/:id
@@ -79,6 +107,11 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
         )
       );
     }
+
+     //Make sure user is bootcamp owner
+     if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+      return next( new ErrorResponse(`User ${req.user.id} is not authorized to delete this bootcamp`, 401));
+      }
 
     bootcamp.deleteOne();
 
@@ -126,14 +159,14 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
   }
 
   // Make sure user is bootcamp owner
-  // if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
-  //   return next(
-  //     new ErrorResponse(
-  //       `User ${req.params.id} is not authorized to update this bootcamp`,
-  //       401
-  //     )
-  //   );
-  // }
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.params.id} is not authorized to update this bootcamp`,
+        401
+      )
+    );
+  }
 
   if (!req.files) {
     return next(new ErrorResponse(`Please upload a file`, 400));
